% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{a10_ts}
\alias{a10_ts}
\title{Monthly Anti-Diabetic Drug Subsidy in Australia from 1991 to 2008.}
\format{
A time series object with 204 observations:
\describe{
  \item{a10}{A numeric vector containing the monthly anti-diabetic drug subsidy in Australia, recorded from 1991 to 2008.}
}
}
\source{
Data provided by the Australian government.
}
\usage{
data(a10_ts)
}
\description{
The dataset name has been changed to 'a10_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'timeseriesdatasets' package and assists users
in identifying its specific characteristics. The suffix 'ts' identifies the dataset as
a time series, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
