% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{airpass_ts}
\alias{airpass_ts}
\title{Monthly Airline Passenger Numbers from 1949 to 1960.}
\format{
A time series object with 144 observations:
\describe{
  \item{airpass}{A numeric vector containing the monthly airline passenger numbers from 1949 to 1960.}
}
}
\source{
Original data provided by the International Air Transport Association (IATA).
}
\usage{
data(airpass_ts)
}
\description{
The dataset name has been changed to 'airpass_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'timeseriesdatasets' package and assists users
in identifying its specific characteristics. The suffix 'ts' identifies the dataset as
a time series, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
