% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ausbeer_ts}
\alias{ausbeer_ts}
\title{Quarterly Australian Beer Production.}
\format{
A time series object with 218 observations:
\describe{
  \item{ausbeer}{A numeric vector representing the quarterly beer production in Australia, measured in megalitres.}
}
}
\source{
Data provided by the Australian Bureau of Statistics.
}
\usage{
data(ausbeer_ts)
}
\description{
The dataset name has been changed to 'ausbeer_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'timeseriesdatasets' package and assists users
in identifying its specific characteristics. The suffix 'ts' identifies the dataset as
a time series object, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
