% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{auscafe_ts}
\alias{auscafe_ts}
\title{Monthly Expenditure on Eating Out in Australia.}
\format{
A time series object with 426 observations:
\describe{
  \item{auscafe}{A numeric vector representing the monthly expenditure on eating out in Australia, measured in dollars.}
}
}
\source{
Data provided by the Australian Bureau of Statistics.
}
\usage{
data(auscafe_ts)
}
\description{
The dataset name has been changed to 'auscafe_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'timeseriesdatasets' package and assists users
in identifying its specific characteristics. The suffix 'ts' identifies the dataset as
a time series object, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
