% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{beer_ts}
\alias{beer_ts}
\title{Monthly Beer Production.}
\format{
A time series object with 56 observations:
\describe{
  \item{beer}{A numeric vector representing the monthly beer production, measured in megalitres.}
}
}
\source{
Data provided by the Australian Bureau of Statistics.
}
\usage{
data(beer_ts)
}
\description{
The dataset name has been changed to 'beer_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' signifies that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the 'timeseriesdatasets'
package and assists users in identifying its specific characteristics. The original content of the dataset has
not been modified in any way.
}
