% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{BJsales_ts}
\alias{BJsales_ts}
\title{Sales Data with Leading Indicator.}
\format{
A time series object with 150 observations:
\describe{
  \item{BJsales}{A numeric vector representing the sales data with a leading indicator, measured in dollars.}
}
}
\source{
Original data provided by Box and Jenkins (1976) in their book "Time Series Analysis: Forecasting and Control."
}
\usage{
data(BJsales_ts)
}
\description{
The dataset name has been changed to 'BJsales_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. This naming convention helps
distinguish this dataset as part of the 'timeseriesdatasets' package and assists users
in identifying its specific characteristics. The suffix 'ts' identifies the dataset as
a time series object, helping to differentiate it from other datasets within the package.
The original content of the dataset has not been modified in any way.
}
