% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{books_mts}
\alias{books_mts}
\title{Sales of Paperback and Hardcover Books.}
\format{
A time series object with 60 observations and 2 series:
\describe{
  \item{Paperback}{A numeric vector representing the monthly sales of paperback books.}
  \item{Hardcover}{A numeric vector representing the monthly sales of hardcover books.}
}
}
\source{
Data provided by the Australian Bureau of Statistics.
}
\usage{
data(books_mts)
}
\description{
The dataset name has been changed to 'books_mts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'mts' signifies that this dataset
is a multivariate time series object. This naming convention helps distinguish this dataset as part of the 'timeseriesdatasets'
package and assists users in identifying its specific characteristics. The original content of the dataset has
not been modified in any way.
}
