% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{bricksq_ts}
\alias{bricksq_ts}
\title{Quarterly Clay Brick Production.}
\format{
A time series object with 155 observations:
\describe{
  \item{bricksq}{A numeric vector representing the quarterly clay brick production, measured in thousands of bricks.}
}
}
\source{
Data provided by the Australian Bureau of Statistics.
}
\usage{
data(bricksq_ts)
}
\description{
The dataset name has been changed to 'bricksq_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' signifies that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the 'timeseriesdatasets'
package and assists users in identifying its specific characteristics. The original content of the dataset has
not been modified in any way.
}
