% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{co2_ts}
\alias{co2_ts}
\title{Mauna Loa Atmospheric CO2 Concentration.}
\format{
A time series object with 468 observations:
\describe{
  \item{co2}{A numeric vector representing the atmospheric CO2 concentration measured at Mauna Loa, Hawaii, in parts per million (ppm).}
}
}
\source{
Data provided by the National Oceanic and Atmospheric Administration (NOAA).
}
\usage{
data(co2_ts)
}
\description{
The dataset name has been changed to 'co2_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
