% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{discoveries_ts}
\alias{discoveries_ts}
\title{Yearly Numbers of Important Discoveries.}
\format{
A time series object with 100 observations:
\describe{
  \item{discoveries}{A numeric vector representing the yearly number of important discoveries.}
}
}
\source{
Data provided by the dataset included in R's `datasets` package.
}
\usage{
data(discoveries_ts)
}
\description{
The dataset name has been changed to 'discoveries_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
