% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{economics_df_ts}
\alias{economics_df_ts}
\title{US Economic Time Series.}
\format{
A time series object (`spec_tbl_df`) with 574 rows and 6 variables:
\describe{
  \item{date}{A Date vector representing the date of the observation.}
  \item{pce}{A numeric vector representing the personal consumption expenditures (PCE) in billions of dollars.}
  \item{pop}{A numeric vector representing the US population in millions.}
  \item{psavert}{A numeric vector representing the personal saving rate as a percentage of disposable income.}
  \item{uempmed}{A numeric vector representing the median duration of unemployment in weeks.}
  \item{unemploy}{A numeric vector representing the number of unemployed individuals in thousands.}
}
}
\source{
Data provided by the Federal Reserve Economic Data (FRED) database.
}
\usage{
data(economics_df_ts)
}
\description{
The dataset name has been changed to 'economics_df_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'df_ts' indicates that this dataset
is a data frame containing time series data. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
