% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{elec_ts}
\alias{elec_ts}
\title{Electricity Production.}
\format{
A time series object with 476 observations:
\describe{
  \item{elec}{A numeric vector representing the monthly electricity production, measured in gigawatt-hours (GWh).}
}
}
\source{
Data provided by the Australian Bureau of Statistics.
}
\usage{
data(elec_ts)
}
\description{
The dataset name has been changed to 'elec_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
