% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{elecdaily_mts}
\alias{elecdaily_mts}
\title{Half-Hourly and Daily Electricity Demand for Victoria, Australia, in 2014.}
\format{
A time series object with 365 observations and 3 variables:
\describe{
  \item{Demand}{A numeric vector representing the half-hourly electricity demand in megawatts (MW).}
  \item{WorkDay}{A binary vector indicating whether the day is a working day (1 = yes, 0 = no).}
  \item{Temperature}{A numeric vector representing the daily average temperature in degrees Celsius.}
}
}
\source{
Data provided by the Australian Energy Market Operator (AEMO).
}
\usage{
data(elecdaily_mts)
}
\description{
The dataset name has been changed to 'elecdaily_mts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'mts' indicates that this dataset
is a multivariate time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
