% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{elecdemand_msts}
\alias{elecdemand_msts}
\title{Half-Hourly and Daily Electricity Demand for Victoria, Australia, in 2014.}
\format{
A time series object with multiple seasonalities and 17,520 observations:
\describe{
  \item{Demand}{A numeric vector representing the half-hourly electricity demand in megawatts (MW).}
  \item{WorkDay}{A binary vector indicating whether the half-hour period is on a working day (1 = yes, 0 = no).}
  \item{Temperature}{A numeric vector representing the half-hourly temperature in degrees Celsius.}
}
}
\source{
Data provided by the Australian Energy Market Operator (AEMO).
}
\usage{
data(elecdemand_msts)
}
\description{
The dataset name has been changed to 'elecdemand_msts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'msts' indicates that this dataset
is a time series object with multiple seasonalities. This naming convention helps distinguish this
dataset as part of the 'timeseriesdatasets' package and assists users in identifying its specific characteristics.
The original content of the dataset has not been modified in any way.
}
