% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{elecequip_ts}
\alias{elecequip_ts}
\title{Electrical Equipment Manufactured in the Euro Area.}
\format{
A time series object with 195 observations:
\describe{
  \item{elecequip}{A numeric vector representing the monthly production of electrical equipment in the Euro area, measured in millions of euros.}
}
}
\source{
Data provided by Eurostat.
}
\usage{
data(elecequip_ts)
}
\description{
The dataset name has been changed to 'elecequip_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
