% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{euretail_ts}
\alias{euretail_ts}
\title{Quarterly Retail Trade in the Euro Area.}
\format{
A time series object with 64 observations:
\describe{
  \item{euretail}{A numeric vector representing the quarterly retail trade in the Euro area, measured in millions of euros.}
}
}
\source{
Data provided by Eurostat.
}
\usage{
data(euretail_ts)
}
\description{
The dataset name has been changed to 'euretail_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
