% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{goog200_ts}
\alias{goog200_ts}
\title{Daily Closing Stock Prices of Google Inc. (200 Days).}
\format{
A time series object with 200 observations:
\describe{
  \item{goog200}{A numeric vector representing the daily closing stock prices of Google Inc., measured in US dollars, for a period of 200 days.}
}
}
\source{
Data provided by Yahoo Finance.
}
\usage{
data(goog200_ts)
}
\description{
The dataset name has been changed to 'goog200_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
