% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{gtemp_land_ts}
\alias{gtemp_land_ts}
\title{Global Mean Land Temperature Deviations (1850-2023).}
\format{
A time series object with 174 observations:
\describe{
  \item{gtemp_land}{A numeric vector representing the global mean land temperature deviations from the long-term average, measured in degrees Celsius.}
}
}
\source{
Data provided by the National Oceanic and Atmospheric Administration (NOAA) and the Hadley Centre.
}
\usage{
data(gtemp_land_ts)
}
\description{
The dataset name has been changed to 'gtemp_land_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
