% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{h02_ts}
\alias{h02_ts}
\title{Monthly Corticosteroid Drug Subsidy in Australia (1991-2008).}
\format{
A time series object with 204 observations:
\describe{
  \item{h02}{A numeric vector representing the monthly corticosteroid drug subsidy in Australia, measured in Australian dollars, from January 1991 to December 2008.}
}
}
\source{
Data provided by the Australian Government Department of Health.
}
\usage{
data(h02_ts)
}
\description{
The dataset name has been changed to 'h02_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
