% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{hsales2_ts}
\alias{hsales2_ts}
\title{Sales of New One-Family Houses (1987-1996).}
\format{
A time series object with 107 observations:
\describe{
  \item{hsales2}{A numeric vector representing the monthly sales of new one-family houses, measured in thousands of units, from January 1987 to December 1996.}
}
}
\source{
Data provided by the U.S. Census Bureau and the U.S. Department of Housing and Urban Development.
}
\usage{
data(hsales2_ts)
}
\description{
The dataset name has been changed to 'hsales2_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
