% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{hyndsight_ts}
\alias{hyndsight_ts}
\title{Daily Pageviews for the Hyndsight Blog (April 2014 - April 2015).}
\format{
A time series object with 365 observations:
\describe{
  \item{hyndsight}{A numeric vector representing the daily pageviews for the Hyndsight blog, measured as the number of pageviews per day, from April 30, 2014 to April 29, 2015.}
}
}
\source{
Data provided by the author of the Hyndsight blog.
}
\usage{
data(hyndsight_ts)
}
\description{
The dataset name has been changed to 'hyndsight_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
