% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ibm_mts}
\alias{ibm_mts}
\title{IBM Sales and Profit Data.}
\format{
A time series object with 42 observations and 4 variables:
\describe{
  \item{Sales}{A numeric vector representing the monthly sales of IBM, measured in millions of dollars.}
  \item{Profit}{A numeric vector representing the monthly profit of IBM, measured in millions of dollars.}
  \item{FSales}{A numeric vector representing the forecasted sales of IBM.}
  \item{FProfit}{A numeric vector representing the forecasted profit of IBM.}
}
}
\source{
Data provided by the IBM Company.
}
\usage{
data(ibm_mts)
}
\description{
The dataset name has been changed to 'ibm_mts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'mts' indicates that this dataset
is a multivariate time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
