% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ibmclose_ts}
\alias{ibmclose_ts}
\title{Daily Closing Stock Prices of IBM.}
\format{
A time series object with 369 observations:
\describe{
  \item{ibmclose}{A numeric vector representing the daily closing stock price of IBM, measured in US dollars, over a period of 369 days.}
}
}
\source{
Data provided by Yahoo Finance.
}
\usage{
data(ibmclose_ts)
}
\description{
The dataset name has been changed to 'ibmclose_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
