% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{jcars_ts}
\alias{jcars_ts}
\title{Motor Vehicle Production (1947-1989).}
\format{
A time series object with 43 observations:
\describe{
  \item{jcars}{A numeric vector representing the monthly production of motor vehicles, measured in thousands of units, from January 1947 to December 1989.}
}
}
\source{
Data provided by the International Organization of Motor Vehicle Manufacturers (OICA).
}
\usage{
data(jcars_ts)
}
\description{
The dataset name has been changed to 'jcars_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
