% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{JohnsonJohnson_ts}
\alias{JohnsonJohnson_ts}
\title{Quarterly Earnings per Johnson & Johnson Share (1960-1981).}
\format{
A time series object with 84 observations:
\describe{
  \item{JohnsonJohnson}{A numeric vector representing the quarterly earnings per share of Johnson & Johnson, measured in US dollars, from the first quarter of 1960 to the fourth quarter of 1981.}
}
}
\source{
Data provided by the Johnson & Johnson Company.
}
\usage{
data(JohnsonJohnson_ts)
}
\description{
The dataset name has been changed to 'JohnsonJohnson_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
