% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{LakeHuron_ts}
\alias{LakeHuron_ts}
\title{Lake Huron Water Level (1875-1972).}
\format{
A time series object with 98 observations:
\describe{
  \item{LakeHuron}{A numeric vector representing the annual water level of Lake Huron, measured in feet, from 1875 to 1972.}
}
}
\source{
Data provided by the National Oceanic and Atmospheric Administration (NOAA).
}
\usage{
data(LakeHuron_ts)
}
\description{
The dataset name has been changed to 'LakeHuron_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
