% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{ldeaths_ts}
\alias{ldeaths_ts}
\title{Monthly Deaths from Lung Diseases in the UK (1974-1980).}
\format{
A time series object with 72 observations:
\describe{
  \item{ldeaths}{A numeric vector representing the monthly number of deaths from lung diseases in the UK, from 1974 to 1980.}
}
}
\source{
Data collected by the UK Office of Population Censuses and Surveys.
}
\usage{
data(ldeaths_ts)
}
\description{
The dataset name has been changed to 'ldeaths_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
