% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{livestock_ts}
\alias{livestock_ts}
\title{Livestock (Sheep) in Asia, 1961-2007.}
\format{
A time series object with 47 observations:
\describe{
  \item{livestock}{A numeric vector representing the annual number of livestock (sheep) in Asia from 1961 to 2007.}
}
}
\source{
Food and Agriculture Organization of the United Nations (FAO).
}
\usage{
data(livestock_ts)
}
\description{
The dataset name has been changed to 'livestock_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
