% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{marathon_ts}
\alias{marathon_ts}
\title{Boston Marathon Winning Times Since 1897}
\format{
A time series object with 120 observations:
\describe{
  \item{marathon}{A numeric vector representing the winning times (in minutes) of the Boston Marathon from 1897 to 2016.}
}
}
\source{
Boston Athletic Association.
}
\usage{
data(marathon_ts)
}
\description{
The dataset name has been changed to 'marathon_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
