% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{maxtemp_ts}
\alias{maxtemp_ts}
\title{Maximum Annual Temperatures at Moorabbin Airport, Melbourne}
\format{
A time series object with 46 observations:
\describe{
  \item{maxtemp}{A numeric vector representing the maximum annual temperatures (in degrees Celsius) recorded at Moorabbin Airport, Melbourne, from 1971 to 2016.}
}
}
\source{
Australian Bureau of Meteorology.
}
\usage{
data(maxtemp_ts)
}
\description{
The dataset name has been changed to 'maxtemp_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
