% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{mdeaths_ts}
\alias{mdeaths_ts}
\title{Monthly Deaths from Lung Diseases in the UK}
\format{
A time series object of 72 observations on the following:
\describe{
  \item{Time-Series [1:72]}{The number of monthly deaths from lung diseases in the UK.}
}
}
\source{
UK Government Health Department
}
\usage{
data(mdeaths_ts)
}
\description{
The dataset name has been changed to 'mdeaths_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
The data covers the period from 1974 to 1980.
}
