% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{mens400_ts}
\alias{mens400_ts}
\title{Winning Times in Olympic Men's 400m Track Final}
\format{
A time series object of 31 observations on the following:
\describe{
  \item{Time-Series [1:31]}{Winning times (in seconds) in the Olympic men's 400m track final.}
}
}
\source{
International Olympic Committee (IOC)
}
\usage{
data(mens400_ts)
}
\description{
The dataset name has been changed to 'mens400_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
The data covers the period from 1896 to 2016.
}
