% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{milk_ts}
\alias{milk_ts}
\title{Monthly Milk Production per Cow}
\format{
A time series object of 168 observations on the following:
\describe{
  \item{Time-Series [1:168]}{Monthly milk production per cow (in pounds).}
}
}
\source{
U.S. Department of Agriculture
}
\usage{
data(milk_ts)
}
\description{
The dataset name has been changed to 'milk_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
The data covers the period from 1962 to 1976.
}
