% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{nail_ts}
\alias{nail_ts}
\title{Nail Prices}
\format{
A time series object of 197 observations on the following:
\describe{
  \item{Time-Series [1:197]}{Nail prices (in cents per pound).}
}
}
\source{
U.S. Department of Agriculture
}
\usage{
data(nail_ts)
}
\description{
The dataset name has been changed to 'nail_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
The data covers the period from 1800 to 1996.
}
