% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Nile_ts}
\alias{Nile_ts}
\title{Flow of the River Nile}
\format{
A time series object of 100 observations on the following:
\describe{
  \item{Time-Series [1:100]}{Monthly flow of the River Nile (in cubic meters per second).}
}
}
\source{
The dataset is from the British Hydrographic Office and has been digitized by the R package datasets.
}
\usage{
data(Nile_ts)
}
\description{
The dataset name has been changed to 'Nile_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
The data covers the period from 1871 to 1970.
}
