% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{pedestrian_tbl_ts}
\alias{pedestrian_tbl_ts}
\title{Pedestrian Counts in the City of Melbourne}
\format{
A `tbl_ts` object with 66,037 rows and 5 columns:
\describe{
  \item{Sensor}{Identifier for the sensor used to count pedestrians.}
  \item{Date_Time}{Date and time of the pedestrian count (POSIXct format).}
  \item{Date}{Date of the pedestrian count (Date format).}
  \item{Time}{Time of the pedestrian count (Time format).}
  \item{Count}{Number of pedestrians counted.}
}
}
\source{
City of Melbourne, Australia.
}
\usage{
data(pedestrian_tbl_ts)
}
\description{
The dataset name has been changed to 'pedestrian_tbl_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'tbl_ts' indicates that this dataset
is a tibble-based time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains pedestrian count data collected from various sensors located throughout the city of Melbourne. The data includes information on the date and time of each count as well as the number of pedestrians detected.
}
\keyword{datasets}
