% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{qauselec_ts}
\alias{qauselec_ts}
\title{Quarterly Australian Electricity Production}
\format{
A `ts` object with 218 observations:
\describe{
  \item{time}{Quarterly time series from 1956 to 2010.}
  \item{value}{Electricity production (in GWh) in Australia.}
}
}
\source{
Australian Bureau of Statistics.
}
\usage{
data(qauselec_ts)
}
\description{
The dataset name has been changed to 'qauselec_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains quarterly data on electricity production in Australia from 1956 to 2010. The values are measured in gigawatt-hours (GWh).
}
\keyword{datasets}
