% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{qcement_ts}
\alias{qcement_ts}
\title{Quarterly Australian Portland Cement Production}
\format{
A `ts` object with 233 observations:
\describe{
  \item{time}{Quarterly time series from 1956 to 2014.}
  \item{value}{Portland cement production (in thousands of tonnes) in Australia.}
}
}
\source{
Australian Bureau of Statistics.
}
\usage{
data(qcement_ts)
}
\description{
The dataset name has been changed to 'qcement_ts' to avoid confusion with other packages
in the R ecosystem from which datasets have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains quarterly data on Portland cement production in Australia from 1956 to 2014. The values are measured in thousands of tonnes.
}
\keyword{datasets}
