% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesdatasets-package.R
\docType{package}
\name{timeseriesdatasets}
\alias{timeseriesdatasets}
\alias{timeseriesdatasets-package}
\title{timeseriesdatasets: Collection of Time Series Data Sets}
\description{
{
This package provides a wide variety of time series data sets for testing, learning, and research purposes.
}
}
\details{
timeseriesdatasets: Collection of Time Series Data Sets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive collection of time series data sets for use in R.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/timeseriesdatasets_R}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
