% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{uschange_mts}
\alias{uschange_mts}
\title{Growth Rates of Personal Consumption and Personal Income in the USA}
\format{
A `ts` object with 187 observations and 5 variables:
\describe{
  \item{Consumption}{Growth rate of personal consumption.}
  \item{Income}{Growth rate of personal income.}
  \item{Production}{Growth rate of industrial production.}
  \item{Savings}{Growth rate of personal savings.}
  \item{Unemployment}{Growth rate of unemployment.}
}
}
\source{
Federal Reserve Economic Data (FRED).
}
\usage{
data(uschange_mts)
}
\description{
The dataset name has been changed to 'uschange_mts' to avoid confusion with other datasets
in the R ecosystem from which data have been sourced. The suffix 'mts' indicates that this dataset
is a multivariate time series object. This naming convention helps distinguish this dataset
as part of the 'timeseriesdatasets' package and assists users in identifying its specific characteristics.
The original content of the dataset has not been modified in any way.
}
\details{
This dataset contains quarterly growth rates of various economic indicators in the USA from 1970 to 2016.
The variables include rates of personal consumption, personal income, industrial production, personal savings,
and unemployment.
}
\keyword{datasets}
