% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{USgas_ts}
\alias{USgas_ts}
\title{US Monthly Natural Gas Consumption}
\format{
A `ts` object with 238 observations:
\describe{
  \item{time}{Monthly time series from 2000 to 2020.}
  \item{value}{Natural gas consumption in the US (in billion cubic feet).}
}
}
\source{
U.S. Energy Information Administration.
}
\usage{
data(USgas_ts)
}
\description{
The dataset name has been changed to 'USgas_ts' to avoid confusion with other datasets
in the R ecosystem from which data have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics.
The original content of the dataset has not been modified in any way.
}
\details{
This dataset contains monthly data on natural gas consumption in the United States from 2000
to 2020. The values represent consumption in billion cubic feet.
}
\keyword{datasets}
