% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{uspop_ts}
\alias{uspop_ts}
\title{US Census Population Data}
\format{
A `ts` object with 19 observations:
\describe{
  \item{time}{Annual time series from 1790 to 1970.}
  \item{value}{Population recorded by the US Census (in millions).}
}
}
\source{
US Census Bureau.
}
\usage{
data(uspop_ts)
}
\description{
The dataset name has been changed to 'uspop_ts' to avoid confusion with other datasets
in the R ecosystem from which data have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics.
The original content of the dataset has not been modified in any way.
}
\details{
This dataset contains annual population data recorded by the US Census from 1790 to 1970.
The values are recorded in millions.
}
\keyword{datasets}
