% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{wmurders_ts}
\alias{wmurders_ts}
\title{Annual Female Murder Rate in the USA}
\format{
A `ts` object with 55 observations:
\describe{
  \item{time}{Annual time series from 1950 to 2004.}
  \item{value}{Female murder rate per 100,000 standard population in the USA.}
}
}
\source{
United States Federal Bureau of Investigation (FBI).
}
\usage{
data(wmurders_ts)
}
\description{
The dataset name has been changed to 'wmurders_ts' to avoid confusion with other datasets
in the R ecosystem from which data have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains annual data on the female murder rate per 100,000 standard population in the USA from 1950 to 2004.
}
\keyword{datasets}
