% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{woolyrnq_ts}
\alias{woolyrnq_ts}
\title{Quarterly Production of Woollen Yarn in Australia}
\format{
A `ts` object with 119 observations:
\describe{
  \item{time}{Quarterly time series from 1965 to 1994.}
  \item{value}{Quarterly production of woollen yarn in Australia (in millions of dollars).}
}
}
\source{
Australian Bureau of Statistics.
}
\usage{
data(woolyrnq_ts)
}
\description{
The dataset name has been changed to 'woolyrnq_ts' to avoid confusion with other datasets
in the R ecosystem from which data have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains quarterly data on the production of woollen yarn in Australia from 1965 to 1994. The values are measured in millions of dollars.
}
\keyword{datasets}
