% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{WWWusage_ts}
\alias{WWWusage_ts}
\title{Internet Usage per Minute}
\format{
A `ts` object with 100 observations:
\describe{
  \item{time}{Time series from 1 to 100 minutes.}
  \item{value}{Internet usage per minute (in MB).}
}
}
\source{
The dataset is part of the R datasets package.
}
\usage{
data(WWWusage_ts)
}
\description{
The dataset name has been changed to 'WWWusage_ts' to avoid confusion with other datasets
in the R ecosystem from which data have been sourced. The suffix 'ts' indicates that this dataset
is a time series object. This naming convention helps distinguish this dataset as part of the
'timeseriesdatasets' package and assists users in identifying its specific characteristics. The original content
of the dataset has not been modified in any way.
}
\details{
This dataset contains data on internet usage per minute from 1 to 100 minutes. The values are measured in megabytes (MB).
}
\keyword{datasets}
